requirements.txt
streamlit
google-generativeai
pandas
python-docx
pypdf
Pillow